organize directory - remove notion_df/scratch to personal vault, form -> components
