+++ notion_df/component/api/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from abc import abstractmethod, ABCMeta
+from abc import ABCMeta
 
 from notion_df.util.mixin import input_based_cache
 
@@ -15,13 +15,8 @@ def _get_headers(notion_api_key: str):
     }
 
 
-class RequestBuilder(metaclass=ABCMeta):
+class APIRequest(metaclass=ABCMeta):
     ENDPOINT = "https://api.notion.com/v1/"
 
     def __init__(self, notion_api_key: str):
         self.headers = _get_headers(notion_api_key)
-
-    @classmethod
-    @abstractmethod
-    def _get_entrypoint(cls):
-        pass
